requirements.txt
python-telegram-bot
pymongo
requests
openai
flask
